# src/main/java/org/tkit/parameters/rs/internal/ApplicationParameterHistoryRestController.java
package org.tkit.parameters.rs.internal;

import jakarta.enterprise.context.ApplicationScoped;
import jakarta.inject.Inject;
import jakarta.transaction.Transactional;
import jakarta.ws.rs.*;
import jakarta.ws.rs.core.MediaType;
import jakarta.ws.rs.core.Response;

import org.eclipse.microprofile.openapi.annotations.Operation;
import org.eclipse.microprofile.openapi.annotations.enums.SchemaType;
import org.eclipse.microprofile.openapi.annotations.media.Content;
import org.eclipse.microprofile.openapi.annotations.media.Schema;
import org.eclipse.microprofile.openapi.annotations.responses.APIResponse;
import org.eclipse.microprofile.openapi.annotations.tags.Tag;
import org.tkit.parameters.domain.criteria.ApplicationParameterHistorySearchCriteria;
import org.tkit.parameters.domain.daos.ApplicationParameterHistoryDAO;
import org.tkit.parameters.domain.models.ApplicationParameterHistory;
import org.tkit.parameters.rs.internal.dtos.*;
import org.tkit.parameters.rs.internal.mappers.ApplicationParameterInternalMapper;
import org.tkit.quarkus.jpa.daos.PageResult;

@ApplicationScoped
@Tag(name = "internal")
@Path("/histories")
@Produces(MediaType.APPLICATION_JSON)
@Consumes(MediaType.APPLICATION_JSON)
@Transactional(value = Transactional.TxType.NOT_SUPPORTED)
public class ApplicationParameterHistoryRestController {

    @Inject
    ApplicationParameterInternalMapper applicationParameterInternalMapper;

    @Inject
    ApplicationParameterHistoryDAO historyDAO;

    @GET
    @Operation(operationId = "getAllApplicationParametersHistory", description = "Find all parameters history")
    @APIResponse(responseCode = "200", description = "OK", content = @Content(mediaType = MediaType.APPLICATION_JSON, schema = @Schema(implementation = ApplicationParameterHistoryPageResultDTO.class)))
    @APIResponse(responseCode = "400", description = "Bad request")
    @APIResponse(responseCode = "500", description = "Internal Server Error")
    public Response getAllApplicationParametersHistory(@BeanParam ApplicationParameterHistorySearchCriteriaDTO dto,
            @QueryParam("latest") Boolean latest,
            @QueryParam("distinct") Boolean distinct) {
        ApplicationParameterHistorySearchCriteria criteria = applicationParameterInternalMapper.map(dto);
        PageResult<ApplicationParameterHistory> parametersHistories;
        // TODO: fix this
        if (Boolean.TRUE.equals(latest)) {
            parametersHistories = historyDAO.searchOnlyLatestByCriteria(criteria);
        } else if (Boolean.TRUE.equals(distinct)) {
            parametersHistories = historyDAO.searchDistinctByCriteria(criteria);
        } else {
            parametersHistories = historyDAO.searchByCriteria(criteria);
        }
        return Response.ok(applicationParameterInternalMapper.mapHistory(parametersHistories)).build();
    }

    @GET
    @Path("/{id}")
    @Operation(operationId = "getApplicationParametersHistoryById", description = "Find parameters history by Id")
    @APIResponse(responseCode = "200", description = "OK", content = @Content(mediaType = MediaType.APPLICATION_JSON, schema = @Schema(implementation = ApplicationParameterHistoryDTO.class)))
    @APIResponse(responseCode = "400", description = "Bad request")
    @APIResponse(responseCode = "404", description = "Not Found")
    @APIResponse(responseCode = "500", description = "Internal Server Error")
    public Response getApplicationParametersHistoryById(@PathParam("id") String id) {
        ApplicationParameterHistory parameter = historyDAO.findById(id);
        if (parameter == null) {
            return Response.status(Response.Status.NOT_FOUND).build();
        }
        return Response.ok(applicationParameterInternalMapper.mapHistory(parameter)).build();
    }

    @GET
    @Path("/counts")
    @Operation(operationId = "getCountsByCriteria", description = "Get creation dates and counts by criteria")
    @APIResponse(responseCode = "200", description = "OK", content = @Content(mediaType = MediaType.APPLICATION_JSON, schema = @Schema(implementation = ParameterHistoryCountDTO.class, type = SchemaType.ARRAY)))
    @APIResponse(responseCode = "400", description = "Bad request")
    @APIResponse(responseCode = "500", description = "Internal Server Error")
    public Response getCountsByCriteria(@BeanParam ApplicationParameterHistorySearchCriteriaDTO dto) {
        ApplicationParameterHistorySearchCriteria criteria = applicationParameterInternalMapper.map(dto);
        var counts = historyDAO.searchCountsByCriteria(criteria);
        var results = applicationParameterInternalMapper.mapCountList(counts);
        return Response.ok(results).build();
    }

}