# src/main/java/org/tkit/parameters/rs/internal/dtos/ApplicationParameterHistoryPageResultDTO.java
package org.tkit.parameters.rs.internal.dtos;

import io.quarkus.runtime.annotations.RegisterForReflection;

@RegisterForReflection
public class ApplicationParameterHistoryPageResultDTO extends PageResultDTO<ApplicationParameterHistoryDTO> {
}